requirements.txt
python-binance>=0.3
sqlalchemy
schedule
apprise
Flask
gunicorn
flask-cors
flask-socketio
eventlet
python-socketio[client]
diskcache
click
click_default_group
